requirements.txt
numpy==1.21.5
scipy==1.7.3
scikit-learn==1.0.2
scikit-plot==0.3.7
matplotlib==3.5.2
seaborn==0.12.2
pandas==1.3.5
pyyaml==6.0
tqdm==4.64.1
dgl==1.1.0
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
wandb==0.13.7
networkx==2.2
torch-geometric==2.2.0
tensorboard==2.6.0
tensorboardx==2.2